// src/types/types.ts
export enum ApplicationCommandCategory {
  Core = 'Core',
  Forge = 'Forge',
  Dev = 'Dev',
}
// ...
export enum TimestampStyle {
  /**	16:20 */
  ShortTime = 't',
  /**	16:20:30 */
  MediumTime = 'T',
  /**	20/04/2021 */
  ShortDate = 'd',
  /**	April 20, 2021 */
  LongDate = 'D',
  /**	April 20, 2021 at 16:20 */
  LongDateShortTime = 'f',
  /**	Tuesday, April 20, 2021 at 16:20 */
  FullDateShortTime = 'F',
  /**	20/04/2021, 16:20 */
  ShortDateShortTime = 's',
  /**	20/04/2021, 16:20:30 */
  ShortDateMediumTime = 'S',
  /**	4 years ago */
  RelativeTime = 'R',
}

export enum HighlightStyle {
  Bold = 'bold',
  Compact = 'compact',
  Default = 'default',
}

// ========== INTERFACES ==========

export interface Stack {
  reject(res?: unknown): any;
  resolve(val: unknown): any;
  callback: () => any;
}

export interface Details {
  category: ApplicationCommandCategory;
  summary?: string;
  usage?: string;
  examples?: string[];
  cooldown?: number;
}

export interface CommandPermission {
  author: PermissionStrings[];
  client: PermissionStrings[];
}

export interface Precondition {
  run(context: any): boolean | Promise<boolean>;
  fail(context: any): void;
}
// ...
export interface ApplicationCommand<
  T extends ApplicationCommandOptions = ApplicationCommandOptions,
> extends CreateApplicationCommand {
  details: Details;
  preconditions?: Precondition;
  permissions?: CommandPermission;
  options?: T;
  acknowledge?: boolean;
  ephemeral?: boolean;
  dev?: boolean;
  run: (bot: CustomBot, interaction: Interaction, options: GetApplicationCommandOptions<T>) => any;
  autocomplete?: (bot: CustomBot, interaction: Interaction, options: GetApplicationCommandOptions<T>) => any;
}
// ...
export type ApplicationCommandOption = Camelize<DiscordApplicationCommandOption>;
export type ApplicationCommandOptions = ApplicationCommandOption[];
// ...
export type Interaction = typeof bot.transformers.$inferredTypes.interaction;
export type Message = typeof bot.transformers.$inferredTypes.message;
export type Guild = typeof bot.transformers.$inferredTypes.guild;

export type CollectorType<T> = ReturnType<typeof createCollector<T>>;
// ...
export type BuildOptions<T extends ApplicationCommandOptions | undefined> = {
  [Prop in keyof Omit<T, keyof unknown[]> as GetOptionName<T[Prop]>]: GetOptionValue<T[Prop]>;
};

export type GetApplicationCommandOptions<T extends ApplicationCommandOptions> = T extends ApplicationCommandOptions
  ? { [Prop in keyof BuildOptions<T> as Prop]: BuildOptions<T>[Prop] }
  : never;